修改了UserService在Manager里的服务接口 添加了PO

+++ financeCity/src/main/java/edu/nju/service/PushService/PushServiceImpl.java
@@ -9,8 +9,9 @@ import edu.nju.service.POJO.PushContent;
  */
 public class PushServiceImpl extends BaseFunctionService implements PushService {
 
+
     @Override
-    public PushContent updatePushContent(Integer contentID) {
+    public PushContent updatePushContent(Long contentID) {
         return null;
     }
